playground: add presets and make default blank
The default context and inputw were only intended to give an idea of
the syntax, and having presets for interesting examples gives more room
to do that in a way that doesn't pollute the default page.

This also resolves a request from Ben to add a "clear" button, by making
it one of the presets.

These presets are links and navigate the page, so users can use the
browser back button to undo clear/presetting.

Signed-off-by: Scott Linder <Scott.Linder@amd.com>