playground: clean up html generation
The Dom_html library bundled with js_of_ocaml is actually plenty to
build the bits of HTML we currently need, so there was no need to
add another dependency.

All of the `html_of_*` functions are now `elem_of_*` and return a single
Dom_html.htmlElement to keep things simple.

In a couple places this requires some more <span>s than before (to
bundle up multiple children to return as a single element), but the
resulting HTML just maps more closely to the functions which doesn't
seem like a bad thing.

Also make parsing error messages note whether they apply to parsing the
context or parsing the input.

Signed-off-by: Scott Linder <Scott.Linder@amd.com>